feat: Add new icons and improve header/footer navigation
Introduces a new `GearIcon` for potential future use and replaces `ChartBarIcon` with `BullseyeIcon` in the Services section for better semantic representation.

Enhances the header and footer logo links to provide a smooth scrolling experience to the top of the page when on the landing page. This improves user navigation and reinforces the brand's polished feel.

Updates portfolio section text for better community focus.
Changes `Bespoke Website Design` to `Custom Website Design` and `Robust E-commerce Stores` to `E-commerce Stores That Sell` for clearer value proposition in the Services section.